+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## [next]
+
+- BREAKING: `user` will be returned when signing up with auto confirm off ([#63](https://github.com/supabase-community/gotrue-dart/pull/63))
+
 ## [0.1.6]
 
 - fix: fetch the user, if missing, on `/verify` ([#29](https://github.com/supabase-community/gotrue-dart/issues/29))
-- feat: add JWT headers when refreshing token  ([#53](https://github.com/supabase-community/gotrue-dart/pull/53))
-- feat: add `signInWithOpenIDConnect`  ([#61](https://github.com/supabase-community/gotrue-dart/pull/61))
+- feat: add JWT headers when refreshing token ([#53](https://github.com/supabase-community/gotrue-dart/pull/53))
+- feat: add `signInWithOpenIDConnect` ([#61](https://github.com/supabase-community/gotrue-dart/pull/61))
 
 ## [0.1.5]